+++ containers/daap-python-base/CHANGELOG.md
@@ -9,6 +9,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [4.0.0] - 2023-10-05
+
+### Changed
+
+- refactored `data_product_metadata.py` to have a base class and
+  two child classed for data product schema and metadata.
+- removed `DataProductConfig.metadata_spec_path()` and
+  `DataProductConfig.metadata_spec_prefix()` as are redundant now
+  `data_platform_paths.specification_path()` and
+  `data_platform_paths.specification_prefix()` exists in
+  `data_platform_paths`
+
+### Added
+
+- 403 error status code to `data_platform_api_responses.py`
+- `get_new_version` function to `data_platform_paths``
+
 ## [3.3.1] - 2023-10-11
 
 ### Changed